added auth0, weather api base
